% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crtb_lm_sim_runtimes}
\alias{crtb_lm_sim_runtimes}
\title{crtb_lm_sim time elapsed data}
\format{
\subsection{\code{crtb_lm_sim_runtimes}}{

A data frame with 39 rows and 7 columns:
\describe{
\item{instance}{AWS EC2 instance type}
\item{iterations}{Number of simulation iterations}
\item{B}{Number of bootstrap samples}
\item{user}{"CPU time spent by the current process"}
\item{system}{"CPU time spent by the kernel (the operating system) on behalf
of the current process"}
\item{elapsed}{"wall clock time taken to execute the function...plus some
benchmarking code wrapping it"}
\item{Notes}{Notes}
}
}
}
\usage{
crtb_lm_sim_runtimes
}
\description{
This data set explores how long it takes to run the crtb_lm_sim() function on
different AWS EC2 instances under different numbers of simulation iterations
and bootstrap samples.
}
\details{
The quoted run time definitions were taken from:
\url{https://www.r-bloggers.com/2017/05/5-ways-to-measure-running-time-of-r-code/}
}
\keyword{datasets}
