% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtb.R
\name{crtb_np}
\alias{crtb_np}
\title{Complementary Resampling of Tags in Blocks (Non-Pooled Version)}
\usage{
crtb_np(dat, tie_thresh = 0.5, replace = TRUE, sample_fun = NULL)
}
\arguments{
\item{dat}{A data frame containing the data to resample.}

\item{tie_thresh}{Numeric; a threshold between 0 and 1 to decide if there are
too many ties in the resampled data. If the proportion of unique resampled
tags is less than \code{tie_thresh}, the function will return \code{NULL}
for that group and issue a warning. Default is \code{0.5}.}

\item{replace}{Logical; if \code{TRUE} (default), resampling is done with
replacement.}

\item{sample_fun}{A user-defined function for custom sampling. This function
should accept a vector of tags (the data to be resampled) and return a
resampled vector of the same length. If \code{NULL} (default), the standard
\code{sample} function is used with the \code{replace} argument.}
}
\value{
A list containing two elements:
\item{crdat}{The complementary resampled data.}
\item{ordat}{The original resampled data.}
}
\description{
Performs complementary resampling of tags in blocks (crtb) on the provided data
without pooling the groups. This function applies the resampling procedure
separately to each group in the data.
}
\details{
The function applies the complementary resampling method implemented in
\code{\link{crtb_p}} to each group separately, without pooling the data.
}
\keyword{internal}
