% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crtb.R
\name{crtb_p}
\alias{crtb_p}
\title{Complementary Resampling of Tags in Blocks (Pooled Version)}
\usage{
crtb_p(
  dat,
  rowwise = TRUE,
  tie_thresh = 0.5,
  replace = TRUE,
  sample_fun = NULL
)
}
\arguments{
\item{dat}{A dataframe or vector containing the data to resample.}

\item{rowwise}{Logical; if \code{TRUE} (default), tagging is done row-wise
across groups. If \code{FALSE}, tagging is done column-wise within each group.}

\item{tie_thresh}{Numeric; a threshold between 0 and 1 to decide if there are
too many ties in the resampled data. If the proportion of unique resampled tags
is less than \code{tie_thresh}, the function will return \code{NULL} and issue
a warning. Default is \code{0.5}.}

\item{replace}{Logical; if \code{TRUE} (default), resampling is done with
replacement.}

\item{sample_fun}{A user-defined function for custom sampling. This function
should accept a vector of tags (the data to be resampled) and return a
resampled vector of the same length. If \code{NULL} (default), the standard
\code{sample} function is used with the \code{replace} argument.}
}
\value{
A list containing two elements:
\item{crdat}{The complementary resampled data.}
\item{ordat}{The original resampled data.}
}
\description{
Performs complementary resampling of tags in blocks (crtb) on the provided dataset,
pooling all groups together.
}
\details{
The function operates by assigning unique tags to each observation in the data.
It then resamples these tags (with or without replacement) and rearranges the
data into blocks of unique tags and their complements.
}
\keyword{internal}
