% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{type1_2var_cor_n30_wide_skew_sim}
\alias{type1_2var_cor_n30_wide_skew_sim}
\title{Simulation Results: type1_2var_cor_n30_wide_skew_sim}
\format{
\subsection{\code{type1_2var_cor_n30_wide_skew_sim}}{

A data frame with 2 rows and 7 columns:
\describe{
\item{results}{Bootstrap method}
\item{avg_bias}{Average bias across simulation iterations}
\item{avg_se}{Standard error across simulation iterations}
\item{coverage}{Proportion of confidence intervals that include true parameter}
\item{ci_width}{Average confidence interval width}
\item{converged}{Proportion of converged crtb}
\item{rejection}{Rejection rate}
}
}
}
\usage{
type1_2var_cor_n30_wide_skew_sim
}
\description{
Results from simulation study
}
\keyword{datasets}
